requirements.txt
aiohttp
alembic
async-timeout
attrs
cython
discord.py
ffmpeg
Mako
mcstatus
MarkupSafe
moviepy
multidict
numpy
#psycopg2
pynacl
python-dateutil
python-editor
pytube
pyyaml
six
SQLAlchemy
websockets
wheel
yarl
youtube_dl

